+++ pkg/analysis_server/lib/src/cider/rename.dart
@@ -74,7 +74,7 @@ class CiderRenameComputer {
     if (node == null || element == null) {
       return null;
     }
-    if (element.source != null && element.source!.isInSystemLibrary) {
+    if (element.source != null && element.source!.uri.isScheme('dart')) {
       return null;
     }
     if (element is MethodElement && element.isOperator) {