+++ pkg/analyzer/test/src/dart/sdk/sdk_test.dart
@@ -139,7 +139,6 @@ class FolderBasedDartSdkTest with ResourceProviderMixin {
         .getChildAssumingFile("core.dart")
         .toUri())!;
     expect(source, isNotNull);
-    expect(source.isInSystemLibrary, isTrue);
     expect(source.uri.toString(), "dart:core");
   }
 
@@ -150,7 +149,6 @@ class FolderBasedDartSdkTest with ResourceProviderMixin {
     File file = dirDartium.getChildAssumingFile("html_dart2js.dart");
     var source = sdk.fromFileUri(file.toUri())!;
     expect(source, isNotNull);
-    expect(source.isInSystemLibrary, isTrue);
     expect(source.uri.toString(), "dart:html");
   }
 
@@ -161,7 +159,6 @@ class FolderBasedDartSdkTest with ResourceProviderMixin {
     File file = dirCommon.getChildAssumingFile("html_common_dart2js.dart");
     var source = sdk.fromFileUri(file.toUri())!;
     expect(source, isNotNull);
-    expect(source.isInSystemLibrary, isTrue);
     expect(source.uri.toString(), "dart:html_common");
   }
 
@@ -172,7 +169,6 @@ class FolderBasedDartSdkTest with ResourceProviderMixin {
         .getChildAssumingFile("num.dart")
         .toUri())!;
     expect(source, isNotNull);
-    expect(source.isInSystemLibrary, isTrue);
     expect(source.uri.toString(), "dart:core/num.dart");
   }